Revert "moved get_nwb_file to common_nwbfile.py"
This reverts commit 6958a9b9ade1c11f8f2f887f9e4034dcf179712d.
