+++ src/spyglass/common/common_behav.py
@@ -5,11 +5,11 @@ import pynwb
 
 from .common_ephys import Raw  # noqa: F401
 from .common_interval import IntervalList, interval_list_contains
-from .common_nwbfile import Nwbfile, get_nwb_file
+from .common_nwbfile import Nwbfile
 from .common_session import Session  # noqa: F401
 from .common_task import TaskEpoch
 from .dj_helper_fn import fetch_nwb
-from .nwb_helper_fn import get_all_spatial_series, get_data_interface
+from .nwb_helper_fn import get_all_spatial_series, get_data_interface, get_nwb_file
 
 schema = dj.schema('common_behav')
