+++ src/spyglass/common/common_dio.py
@@ -3,9 +3,9 @@ import pynwb
 
 from .common_ephys import Raw
 from .common_interval import IntervalList  # noqa: F401
-from .common_nwbfile import Nwbfile, get_nwb_file
+from .common_nwbfile import Nwbfile
 from .common_session import Session  # noqa: F401
-from .nwb_helper_fn import get_data_interface
+from .nwb_helper_fn import get_data_interface, get_nwb_file
 from .dj_helper_fn import fetch_nwb  # dj_replace
 
 schema = dj.schema('common_dio')
